chore: move /linktree to new arch and fix parsing
